requirements.txt
fastapi
uvicorn
crewai
langgraph
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
tiktoken